ov006: promote dScMgTeresa_c into a single C++ TU (48 of 81)
48 one-function shards become src/actors/dScMgTeresa_c.cpp, one delink
entry, one manifest entry. verify 48/48 MATCH, objisolate clean,
emission order ROM-ascending; linkcheck (pre-promotion, the only usable
route for a text-only TU) SCRATCH-LINK-VERIFIED: 48 LICENSED / 0 HOMELESS
at [4b/8], the whole .text range 0x0211cbd0..0x0211e72c IDENTICAL,
106/106 modules exact, and a full ROM whose sha256 matches the stock
control byte for byte.

WHY 48 AND NOT 81. tu_map calls 0x0211cbd0..0x021207a8 one contiguous run
of 81 functions. A sourceless hole splits it: ROM ordinal 48,
func_ov006_0211e72c, has a symbols.txt row but no src/ file and no
delinks.txt entry at all, so the cartridge's own bytes cover it. It is a
banked near-miss whose floor of 26 divergences four different models have
reconfirmed. Nothing in this tree can express a .text claim with a hole,
so the run is licensed as one of its two contiguous sides, and this is
the larger by member count -- 48 below against 32 above. The upper side
is larger in bytes but carries the run's other unmatched member,
_ZN13dScMgTeresa_c13OnYoshiTryEatEi, which would split it again to 28+3.
(The queue's shard_count of 80 counts source FILES; the run has 81
members and one of them has no file.)

THE DESTRUCTOR IS OUT OF LINE, AND THE RULE SAYING IT CANNOT BE IS
CONDITIONAL. The ROM orders this class D1 (0x0211cbd0) below D0
(0x0211cbf4). The received rule is that only an inline in-class
destructor gives D1-then-D0 and an out-of-line one gives D2/D0/D1. That
holds only while codegen is deferred. Measured both ways on the real
header with the pinned 2004/b56: out-of-line + deferred emits sections
D2, D0, D1; out-of-line + `#pragma defer_codegen off` emits D1, D0, D2 --
the cartridge's order, with the homeless D2 trailing where production
isolation discards it. The inline form is not usable here at all, for a
separate reason: it moves the key function to InitResources at
0x021203fc, on the far side of the hole, after which this TU emits no
vtable, nothing odr-uses the inline destructor and neither variant is
emitted (probe TU: one section, no destructor). include/dScMgTeresa_c.h
is therefore UNCHANGED, which is also what the 32 still-enrolled shards
above the hole need.

Owning the key function means the whole five-level chain's typeinfo rides
along: compiler_only_output carries 12 rows, 11 deadstrip-data (2 x 5
levels + the vtable, every one with a configured ROM home) plus the
homeless D2 as plain deadstrip. externalized_output is EMPTY.

THE THREE PRAGMAS ON THIS SIDE ARE REAL, AND BOTH CONTROLS WERE RUN.
func_ov006_0211d924 and func_ov006_0211db7c want opt_common_subs off and
func_ov006_0211e220 wants opt_strength_reduction off. Delete all three
outright and it is 45/48, with the DIFFs on exactly those three members
(105, 999, 999 words) -- so they are not belt-and-braces. Keep the
push/pop brackets but delete only `defer_codegen off`, source order
unchanged, and it is 45/48 again AND the emission order inverts to "47
ordinal pair(s) NOT in ROM order". One pragma buys the bytes and the
order together, on the opt_common_subs and opt_strength_reduction
families, reconfirming what ov006/dScMgHanachan_c settled.

Reconciliation: three legacy files each spelled their own `struct C` and
the three disagree offset-for-offset, so they are renamed apart
(TeresaPmfA, TeresaPmfB, RowArray) rather than merged; `struct Elem` had
the same collision and splits into SlotElem and HudElem. decl_common.h is
included and matched -- it already declares nine of this TU's members,
seven as void* where the legacy file used char*, so those definitions
take the header's type and a one-line cast alias.

The ratchet moves the SET, not the count: 2681 -> 2681, the two shard
rows that were in the CONVERTED baseline replaced by the two member rows
on the promoted file, with two correct removal records.
