+++ config/arm9/overlays/ov006/delinks.txt
@@ -4813,197 +4813,9 @@ src/actors/dScMgSound_c.cpp:
     complete
     .text start:0x02119904 end:0x0211cbd0
 
-src/_ZN13dScMgTeresa_cD1Ev.cpp:
+src/actors/dScMgTeresa_c.cpp:
     complete
-    .text start:0x0211cbd0 end:0x0211cbf4
-
-src/_ZN13dScMgTeresa_cD0Ev.cpp:
-    complete
-    .text start:0x0211cbf4 end:0x0211cc2c
-
-src/func_ov006_0211cc2c.c:
-    complete
-    .text start:0x0211cc2c end:0x0211cc90
-
-src/func_ov006_0211cc90.c:
-    complete
-    .text start:0x0211cc90 end:0x0211cca8
-
-src/func_ov006_0211cca8.c:
-    complete
-    .text start:0x0211cca8 end:0x0211cd24
-
-src/func_ov006_0211cd24.c:
-    complete
-    .text start:0x0211cd24 end:0x0211ce90
-
-src/func_ov006_0211ce90.c:
-    complete
-    .text start:0x0211ce90 end:0x0211ce94
-
-src/func_ov006_0211ce94.c:
-    complete
-    .text start:0x0211ce94 end:0x0211cef4
-
-src/func_ov006_0211cef4.c:
-    complete
-    .text start:0x0211cef4 end:0x0211d018
-
-src/func_ov006_0211d018.c:
-    complete
-    .text start:0x0211d018 end:0x0211d0f8
-
-src/func_ov006_0211d0f8.c:
-    complete
-    .text start:0x0211d0f8 end:0x0211d224
-
-src/func_ov006_0211d224.c:
-    complete
-    .text start:0x0211d224 end:0x0211d368
-
-src/func_ov006_0211d368.c:
-    complete
-    .text start:0x0211d368 end:0x0211d4e8
-
-src/func_ov006_0211d4e8.c:
-    complete
-    .text start:0x0211d4e8 end:0x0211d5a8
-
-src/func_ov006_0211d5a8.cpp:
-    complete
-    .text start:0x0211d5a8 end:0x0211d608
-
-src/func_ov006_0211d608.c:
-    complete
-    .text start:0x0211d608 end:0x0211d688
-
-src/func_ov006_0211d688.c:
-    complete
-    .text start:0x0211d688 end:0x0211d69c
-
-src/func_ov006_0211d69c.c:
-    complete
-    .text start:0x0211d69c end:0x0211d75c
-
-src/func_ov006_0211d75c.c:
-    complete
-    .text start:0x0211d75c end:0x0211d7b0
-
-src/func_ov006_0211d7b0.c:
-    complete
-    .text start:0x0211d7b0 end:0x0211d7b4
-
-src/func_ov006_0211d7b4.c:
-    complete
-    .text start:0x0211d7b4 end:0x0211d7d8
-
-src/func_ov006_0211d7d8.c:
-    complete
-    .text start:0x0211d7d8 end:0x0211d7ec
-
-src/func_ov006_0211d7ec.c:
-    complete
-    .text start:0x0211d7ec end:0x0211d86c
-
-src/func_ov006_0211d86c.cpp:
-    complete
-    .text start:0x0211d86c end:0x0211d924
-
-src/func_ov006_0211d924.c:
-    complete
-    .text start:0x0211d924 end:0x0211dad0
-
-src/func_ov006_0211dad0.c:
-    complete
-    .text start:0x0211dad0 end:0x0211db7c
-
-src/func_ov006_0211db7c.c:
-    complete
-    .text start:0x0211db7c end:0x0211dce0
-
-src/func_ov006_0211dce0.c:
-    complete
-    .text start:0x0211dce0 end:0x0211dd0c
-
-src/func_ov006_0211dd0c.cpp:
-    complete
-    .text start:0x0211dd0c end:0x0211dd6c
-
-src/func_ov006_0211dd6c.c:
-    complete
-    .text start:0x0211dd6c end:0x0211ddb8
-
-src/func_ov006_0211ddb8.c:
-    complete
-    .text start:0x0211ddb8 end:0x0211ddcc
-
-src/func_ov006_0211ddcc.c:
-    complete
-    .text start:0x0211ddcc end:0x0211de54
-
-src/func_ov006_0211de54.c:
-    complete
-    .text start:0x0211de54 end:0x0211de7c
-
-src/func_ov006_0211de7c.c:
-    complete
-    .text start:0x0211de7c end:0x0211dec0
-
-src/func_ov006_0211dec0.c:
-    complete
-    .text start:0x0211dec0 end:0x0211e020
-
-src/func_ov006_0211e020.c:
-    complete
-    .text start:0x0211e020 end:0x0211e0c8
-
-src/func_ov006_0211e0c8.cpp:
-    complete
-    .text start:0x0211e0c8 end:0x0211e118
-
-src/func_ov006_0211e118.c:
-    complete
-    .text start:0x0211e118 end:0x0211e184
-
-src/func_ov006_0211e184.c:
-    complete
-    .text start:0x0211e184 end:0x0211e220
-
-src/func_ov006_0211e220.c:
-    complete
-    .text start:0x0211e220 end:0x0211e29c
-
-src/func_ov006_0211e29c.c:
-    complete
-    .text start:0x0211e29c end:0x0211e318
-
-src/func_ov006_0211e318.c:
-    complete
-    .text start:0x0211e318 end:0x0211e3e0
-
-src/func_ov006_0211e3e0.c:
-    complete
-    .text start:0x0211e3e0 end:0x0211e460
-
-src/func_ov006_0211e460.c:
-    complete
-    .text start:0x0211e460 end:0x0211e4e0
-
-src/func_ov006_0211e4e0.c:
-    complete
-    .text start:0x0211e4e0 end:0x0211e55c
-
-src/func_ov006_0211e55c.c:
-    complete
-    .text start:0x0211e55c end:0x0211e5cc
-
-src/func_ov006_0211e5cc.c:
-    complete
-    .text start:0x0211e5cc end:0x0211e658
-
-src/func_ov006_0211e658.c:
-    complete
-    .text start:0x0211e658 end:0x0211e72c
+    .text start:0x0211cbd0 end:0x0211e72c
 
 src/func_ov006_0211e7d8.c:
     complete
