+++ config/converted-backslide-exceptions.jsonl
@@ -344,3 +344,5 @@
 {"path": "src/_ZN12dScMgSound_c6RenderEv.cpp", "reason": "ov006: promote dScMgSound_c into one 82-function C++ TU (src/actors/dScMgSound_c.cpp); the 82 one-function shards are absorbed, not lost -- linkcheck scratch-link-verified, 106/106 modules exact, ROM sha256 identical to stock"}
 {"path": "src/_ZN12dScMgSound_cD0Ev.cpp", "reason": "ov006: promote dScMgSound_c into one 82-function C++ TU (src/actors/dScMgSound_c.cpp); the 82 one-function shards are absorbed, not lost -- linkcheck scratch-link-verified, 106/106 modules exact, ROM sha256 identical to stock"}
 {"path": "src/_ZN12dScMgSound_cD1Ev.cpp", "reason": "ov006: promote dScMgSound_c into one 82-function C++ TU (src/actors/dScMgSound_c.cpp); the 82 one-function shards are absorbed, not lost -- linkcheck scratch-link-verified, 106/106 modules exact, ROM sha256 identical to stock"}
+{"path": "src/_ZN13dScMgTeresa_cD0Ev.cpp", "reason": "ov006/dScMgTeresa_c promotion: 48 one-function shards absorbed into src/actors/dScMgTeresa_c.cpp (ROM ordinals 0..47 of the 0x0211cbd0..0x021207a8 run; the other 33 stay enrolled behind a sourceless hole)"}
+{"path": "src/_ZN13dScMgTeresa_cD1Ev.cpp", "reason": "ov006/dScMgTeresa_c promotion: 48 one-function shards absorbed into src/actors/dScMgTeresa_c.cpp (ROM ordinals 0..47 of the 0x0211cbd0..0x021207a8 run; the other 33 stay enrolled behind a sourceless hole)"}
