+++ config/converted-baseline.json
@@ -704,8 +704,6 @@
     "src/_ZN13UpDownLiftBbhD0Ev.cpp",
     "src/_ZN13UpDownLiftBbhD1Ev.cpp",
     "src/_ZN13dScMgTeresa_c6RenderEv.cpp",
-    "src/_ZN13dScMgTeresa_cD0Ev.cpp",
-    "src/_ZN13dScMgTeresa_cD1Ev.cpp",
     "src/_ZN13daObjDorifu_c16CleanupResourcesEP20daObjDorifuResources.cpp",
     "src/_ZN13daObjDorifu_c6RenderEv.cpp",
     "src/_ZN13daObjDorifu_cD0Ev.cpp",
@@ -2144,6 +2142,8 @@
     "src/actors/dScMgSound_c.cpp#_ZN12dScMgSound_c6RenderEv",
     "src/actors/dScMgSound_c.cpp#_ZN12dScMgSound_cD0Ev",
     "src/actors/dScMgSound_c.cpp#_ZN12dScMgSound_cD1Ev",
+    "src/actors/dScMgTeresa_c.cpp#_ZN13dScMgTeresa_cD0Ev",
+    "src/actors/dScMgTeresa_c.cpp#_ZN13dScMgTeresa_cD1Ev",
     "src/actors/daDgr_c.cpp#_ZN7daDgr_c16CleanupResourcesEv",
     "src/actors/daDgr_c.cpp#_ZN7daDgr_c6RenderEv",
     "src/actors/daDgr_c.cpp#_ZN7daDgr_cD0Ev",
