--- src/_ZN13dScMgTeresa_cD0Ev.cpp
@@ -1,9 +1,0 @@
-//cpp
-// @symbol _ZN13dScMgTeresa_cD0Ev
-#include "dScMgTeresa_c.h"
-/* dScMgTeresa_c::~dScMgTeresa_c (D0, deleting destructor) -- dScMgBase_c's
-   own operator delete (its immediate base) covers this; no per-class copy
-   needed. */
-dScMgTeresa_c::~dScMgTeresa_c()
-{
-}
