--- src/_ZN13dScMgTeresa_cD1Ev.cpp
@@ -1,10 +1,0 @@
-//cpp
-// @symbol _ZN13dScMgTeresa_cD1Ev
-#include "dScMgTeresa_c.h"
-/* dScMgTeresa_c::~dScMgTeresa_c (D1, complete-object destructor) -- no
-   members need explicit destruction, so an empty body relies entirely on
-   the compiler-generated own-vtable-write + base-D2-call, same as
-   dScMgCurling_c's. */
-dScMgTeresa_c::~dScMgTeresa_c()
-{
-}
