--- src/func_ov006_0211ce90.c
@@ -1,3 +1,0 @@
-void func_ov006_0211ce90(void)
-{
-}
