--- src/func_ov006_0211d688.c
@@ -1,5 +1,0 @@
-void func_ov006_0211d688(char *p)
-{
-    *(char *)(p + 0x4be0) = 0;
-    *(char *)(p + 0x4be2) = 0;
-}
