--- src/func_ov006_0211d7b0.c
@@ -1,3 +1,0 @@
-void func_ov006_0211d7b0(void)
-{
-}
