--- src/func_ov006_0211d7b4.c
@@ -1,7 +1,0 @@
-void func_ov006_0211d7b4(char *p)
-{
-    *(int *)(p + 0x4bc0) = 327680;
-    *(int *)(p + 0x4bc4) = 262144;
-    *(char *)(p + 0x4bc8) = 1;
-    *(char *)(p + 0x4bc9) = 1;
-}
