--- src/func_ov006_0211d7d8.c
@@ -1,5 +1,0 @@
-void func_ov006_0211d7d8(char *p)
-{
-    *(char *)(p + 0x4bc8) = 0;
-    *(char *)(p + 0x4bc9) = 0;
-}
