--- src/func_ov006_0211dd6c.c
@@ -1,15 +1,0 @@
-void func_ov006_0211dd6c(char *p)
-{
-    *(int *)(p + 0x4ba0) = 655360;
-    *(int *)(p + 0x4ba4) = 393216;
-    *(char *)(p + 0x4bbc) = 1;
-    *(char *)(p + 0x4bb9) = 1;
-    *(char *)(p + 0x4bba) = 0;
-    *(char *)(p + 0x4bb8) = 0;
-    *(short *)(p + 0x4bb0) = 0;
-    *(short *)(p + 0x4bb2) = 0;
-    *(short *)(p + 0x4bb4) = 0;
-    *(char *)(p + 0x4bbb) = 0;
-    *(char *)(p + 0x4bbd) = 0;
-    *(char *)(p + 0x4bbe) = 0;
-}
