--- src/func_ov006_0211ddb8.c
@@ -1,5 +1,0 @@
-void func_ov006_0211ddb8(char *p)
-{
-    *(char *)(p + 0x4bb9) = 0;
-    *(char *)(p + 0x4bbc) = 0;
-}
