--- src/func_ov006_0211e29c.c
@@ -1,9 +1,0 @@
-extern void func_ov004_020b2220(int,int,int,int,int,int,int);
-extern int GetGameLanguage(void);
-extern void DrawOamSprite(int,int,int,int);
-extern int* data_ov006_0213f9d0[];
-void func_ov006_0211e29c(unsigned char* c){
-  if(*(unsigned char*)(c+0x4c1b)==0) return;
-  func_ov004_020b2220(0x80,0x60,*(unsigned short*)(c+0x4c14),1,0,0x800,0);
-  DrawOamSprite(data_ov006_0213f9d0[GetGameLanguage()][3],0x80,0x48,0);
-}
